ProductiveBot/5/requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
openai>=1.0.0
pandas>=2.0.0
notion-client>=2.0.0
aiohttp>=3.8.0
httpx>=0.24.0

# ─── добавляем поддержку скачивания с YouTube ───
yt-dlp[crypto]>=2025.6.9      # «[crypto]» та же запись, что и в pip install
browser-cookie3>=0.19.1       # для чтения куков браузера
pycryptodomex>=3.20            # только Windows